requirements.txt
rethinkdb
minio
python-dotenv
